fix: bug fix self type resolution
