+++ crates/cyrusc_cir/src/walk.rs
@@ -4,13 +4,13 @@ use crate::*;
 use cyrusc_abi::mangling::ABINameMangling;
 use cyrusc_ast::{LiteralKind, SelfModifierKind};
 use cyrusc_resolver::symbols::{LocalOrGlobalSymbol, LocalScopeRef, ResolvedMethod, generate_symbol_id};
-use cyrusc_resolver::{Resolver, set_self_modifier_type_in_func_sig, typed_func_decl_from_func_sig};
+use cyrusc_resolver::{Resolver};
 use cyrusc_tast::generics::generic_type::GenericType;
 use cyrusc_tast::generics::monomorph::MonomorphEntry;
 use cyrusc_tast::generics::substitute::{
     substitute_enum_sig, substitute_func_sig, substitute_struct_sig, substitute_union_sig,
 };
-use cyrusc_tast::sigs::{EnumSig, FuncSig, GlobalVarSig, UnionSig};
+use cyrusc_tast::sigs::{EnumSig, FuncSig, GlobalVarSig, UnionSig, set_self_modifier_type_in_func_sig, typed_func_decl_from_func_sig};
 use cyrusc_tast::types::{PlainType, ResolvedSymbol};
 use cyrusc_tast::{ModuleID, ScopeID, SymbolID};
 use cyrusc_tast::{
